Change project structure so that our design follows clean architecture.
