+++ phase0/CRC/GroceryInventory.md
@@ -8,4 +8,4 @@ Variables:
 
 Collaborators:
 * GroceryItem
-* ShoppingCart
+* Controllers.ShoppingCart
