--- phase0/CRC/Store.md
@@ -1,10 +1,0 @@
-Class name: Store(Command line interface)
-
-Responsibilities:
-* The command line interface, the only file that should be run
-* initalize a menu to accept user's input
-
-Variables: None
-
-Collaborators:
-* Menu
