+++ phase0/walkthrough.md
@@ -1,2 +1,2 @@
-A user can run the Store class to start the program, when the program start, the Store class will call a 'main_menu' method in the 'Menu' class which offers the user three choices, create an account, login to an account or just exit. if the user choose to create an account, then the 'Menu' class will create a 'Customer' and store this customer account in a 'AccountManager' class. Then the 'Menu' class will call a 'log_in_menu' method, after login, the user can view a list of 'GroceryItem' stored in 'GroceryInventory'. While viewing the item list, the user can add items they want to purchase to their 'ShoppingCart'. When the user wants to checkout, a 'Order' will be generated and stored in 'OrderArrangeSystem', then a 'DeliveryMatchingSystem' will match a 'DeliveryPeople' to handle the order. At the meantime, the user can see the status of their order through the 'OrderTrackingSystem'. When the deliverymen completed the delivery, they can confirm their delivery in the 'OrderTrackingSystem'.
+A user can run the Store class to start the program, when the program start, the Store class will call a 'main_menu' method in the 'External_Interface.Menu' class which offers the user three choices, create an account, login to an account or just exit. if the user choose to create an account, then the 'External_Interface.Menu' class will create a 'Customer' and store this customer account in a 'Controllers.AccountManager' class. Then the 'External_Interface.Menu' class will call a 'log_in_menu' method, after login, the user can view a list of 'GroceryItem' stored in 'GroceryInventory'. While viewing the item list, the user can add items they want to purchase to their 'Controllers.ShoppingCart'. When the user wants to checkout, a 'Order' will be generated and stored in 'OrderArrangeSystem', then a 'DeliveryMatchingSystem' will match a 'DeliveryPeople' to handle the order. At the meantime, the user can see the status of their order through the 'OrderTrackingSystem'. When the deliverymen completed the delivery, they can confirm their delivery in the 'OrderTrackingSystem'.
