--- src/main/java/Command Line Interface/Store.java
@@ -1,18 +1,0 @@
-import java.io.*;
-
-/** this is the file that should be run when someone wants to launch the program.
- *
- */
-
-public class Store {
-
-
-    public static void main(String[] args) throws IOException {
-
-        Menu m = new Menu();
-        m.main_menu();
-
-    }
-
-
-}
