+++ src/main/java/Controllers/StoreManager.java
@@ -1,0 +1,6 @@
+// this class provide access to 'GroceryInventory' at Controller level
+
+package Controllers;
+
+public class StoreManager {
+}
