+++ src/main/java/Entities/GroceryItem.java
@@ -1,3 +1,6 @@
+package Entities;
+
+
 public class GroceryItem{
