--- src/main/java/Use Cases/Order.java
@@ -1,2 +1,0 @@
-
-public class Order{}
