+++ src/main/java/Use_Case/Account.java
@@ -1,0 +1,6 @@
+package Use_Case;
+
+// this class stores an account for a 'Customer'
+
+public class Account {
+}
